Update 2DPathProblem.txt
The first version had a logic error where the path output was the path of x steps not what we wanted.  This version outputs what we want which is the resultant index position list as a result of taking either step in the x or y axis direction. 

Also the additional goal was addressed in the latter part of the code where two points are now given and output are the unique paths of the flag value.