Update 2DPathProblem.txt
Updated functions to produce the ability to determine unique paths and their associated index values to solve for the additional goal. 

Now, two paths of the flag_value only and of a unique path set will be output if there exists two paths that a. are the flag_value only and b. unique path sets